% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plot.lefse.results}
\alias{plot.lefse.results}
\title{plot.lefse.results}
\usage{
plot.lefse.results()
}
\description{
This function plots the results from Curtis Huttenhower's LeFSe output
}
\examples{
plot.lefse.results() #must have a file called lefse.res in working directory
plot.lefse.results()
}

